requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-tavily
langgraph
yfinance
tavily-python
duckduckgo-search
ddgs
fastapi
uvicorn
python-dotenv